Moved s3 specific module to drivers/s3/
`DriverUtils` provides constants specific to S3 driver implementation and is
only used from there, so moving it there.
